modificaciones en clase Menu

+++ src/Domain/Flower.java
@@ -1,12 +1,10 @@
 package Domain;
 
-
-
 public class Flower extends  Product {
 
     private String color;
 
-    public Flower(String name,float price, int quantity, String color) {
+    public Flower(String name,double price, int quantity, String color) {
         super(name,price, quantity);
         this.color = color;
     }
